requirements.txt
Flask>=3.0
Flask-Login>=0.6,<0.7
Flask-WTF>=1.1
Flask-Migrate>=4.0
SQLAlchemy>=2.0
pyodbc>=4.0
python-dotenv>=1.0
alembic>=1.11
Werkzeug>=3.0
itsdangerous>=2.1
pytz
email-validator
